requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.1
chart-studio==1.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cricbuzz-py==0.1.2
cssselect2==0.4.1
cycler==0.10.0
debugpy==1.4.1
decorator==4.0.10
defusedxml==0.7.1
Django==3.2.5
django-filter==2.4.0
django-paypal==1.1.1
django-sendsms==0.4
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
flower==1.0.0
gunicorn==20.1.0
html5lib==1.1
humanize==3.10.0
idna==3.2
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.1.0
jedi==0.18.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
kombu==5.1.0
Markdown==3.3.4
matplotlib==3.4.2
matplotlib-inline==0.1.2
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.26
mysqlclient==2.0.3
nbformat==5.1.3
numpy==1.21.0
opencv-python==4.5.3.56
pandas==1.3.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
plotly==5.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.5.1
pycparser==2.20
pycricbuzz==2.4
pycryptodome==3.10.1
Pygments==2.1.3
PyJWT==1.7.1
pyparsing==2.4.7
pyphen==0.11.0
pyrsistent==0.18.0
pyshorteners==1.0.1
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.2.1
rabbitmq==0.2.0
requests==2.26.0
retrying==1.3.3
simplegeneric==0.8.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tenacity==8.0.1
tinycss2==1.1.0
tornado==6.1
traitlets==4.3.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==5.3.0
